patch try 2
git-svn-id: svn+ssh://50.28.74.132/home/luminari/svn/luminari/myproject/trunk@4607 dda7f814-ccd3-4591-a6d7-1bd23e5e64af

+++ mobact.c
@@ -412,7 +412,7 @@ void mobile_echos(struct char_data *ch) {
     ECHO_COUNT(ch) = num_elements;
   /*DEBUG*/
   char buf[MAX_INPUT_LENGTH];
-  sprintf(buf, "%d", num_elements);
+  sprintf(buf, "%d", (int)num_elements);
   act(buf, FALSE, ch, 0, 0, TO_ROOM);
   /*DEBUG*/
